+++ SpringBoot/Bank/src/main/java/com/spring/bank/repository/loansRepository.java
@@ -1,9 +1,11 @@
 package com.spring.bank.repository;
 
+import java.math.BigDecimal;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.spring.bank.entities.Loans;
 
-public interface loansRepository extends JpaRepository<Loans, Integer> {
+public interface loansRepository extends JpaRepository<Loans, BigDecimal> {
 
 }
